# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /createUser:
      # binds a127 app logic to a route
    x-swagger-router-controller: userController
    post:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: createUser
      parameters:
        - in: body
          name: user
          description: The user to create.
          schema:
            type: object
            required:
              - name
              - username
              - email
              - password
            properties:
              name:
                type: string
              username:
                type: string
              email:
                type: string
              password:
                type: string
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            type: object
  /login:
        # binds a127 app logic to a route
      x-swagger-router-controller: userController
      post:
        description: Returns 'Hello' to the caller
        # used as the method name of the controller
        operationId: login
        parameters:
          - in: body
            name: user
            description: The user to create.
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
        responses:
          "200":
            description: Success
            schema:
              type: object
          # responses may fall through to errors
          default:
            description: Error
            schema:
              type: object
  /logout:
          # binds a127 app logic to a route
        x-swagger-router-controller: userController
        post:
          description: Returns 'Hello' to the caller
          # used as the method name of the controller
          operationId: logout
          responses:
            "200":
              description: Success
              schema:
                type: object
            # responses may fall through to errors
            default:
              description: Error
              schema:
                type: object
  /deleteUser:
        # binds a127 app logic to a route
      x-swagger-router-controller: userController
      post:
        description: Returns 'Hello' to the caller
        # used as the method name of the controller
        operationId: deleteUser
        parameters:
          - in: body
            name: user
            description: The user to create.
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
        responses:
          "200":
            description: Success
            schema:
              type: object
          # responses may fall through to errors
          default:
            description: Error
            schema:
              type: object
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw

# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string